finagle-core: Add a reverse engineering based test for Drv
Problem

All our tests for Drv rely on generating a distribution via invocation
of a random number generator and trying to make sure that matches with
what the weights of the probability distribution. This pattern is
vulnerable to the inherent fluctuations to randomness and makes it so
our tests are 'seed' dependent. This can result in lowered confidence in
our tests.

Solution

Provide a test that will reverse engineer the alias tables. The test is
no longer dependent on random number generation. As a bit of irony, we
use property based testing to feed this test randomly generated probability
distributions.

Differential Revision: https://phabricator.twitter.biz/D820500