finagle-core: Make WeightedApertureTest a bit less order dependent
Problem

The WeightedApertureTest relies on a Rng random number generator that is global to
all tests. At the same time we have some tests that are expected to be deterministic
in that they're intended to detect unexpected changes in outcome. This can be tricky
if you add a new test that also uses the Rng that will run before the deterministic
tests because using it will mutate it's state and thus the outcome of future tests.

Solution

Turn it into a OneInstancePerTest suite so that at least each test is independent
of the others. This allows you to add and remove tests without disturbing the
deterministic tests.

Differential Revision: https://phabricator.twitter.biz/D822566