finagle-core: Trim an intermediate vector out of Balancer
Problem

Vectors are great but tend to be relatively heavy in terms of allocations.
We make an intermediate vector full of tuples on every serverset update
just to stuff that into a map structure.

Solution

Instead of using the var-arg construct of mutable.HashMap instead just
iterate over the vector and add the elements directly. This skips the
intermediate vector and the tuples which are also not used internally to
the map.

Differential Revision: https://phabricator.twitter.biz/D829401