finagle-core: make NameTreeFactory configurable by custom hash
Problem:
When given a NameTree.Union, the finagle client selects
where to send traffic to fulfill the weighing requested
on a per-request basis. Service owners would like the
ability to apply the random distribution over a property
of requests.

Solution:
Make it possible for users to set a context on a request
with a custom Long. This long is then used as the seed
for the random number generator, ensuring that requests
with the same context id will send traffic to the same
"branch" of the NameTree.Union

JIRA Issues: CSL-11520

Differential Revision: https://phabricator.twitter.biz/D817573