finagle-core: populate annotation for dark traffic in light service
Problem
The dark traffic annotation is only accessible through tracing and
Finagle Local context within a dark service. This behavior can be
demonstrated in the code block below:

```
// light service issues the request,
// it doesn't have the dark traffic annotation
service(request).ensure {
  // we have annotation in `sendDarkTraffic()`,
  // where dark service can see
  sendDarkTraffic()
}
```

Solution
Populate an annotation before sending traffic to dark service so the
annotation is also accessible from the light service.

Furthermore, the methods in AbstractDarkTrafficFilter require a
function of type (Req => Boolean) to determine if dark traffic
should be forwarded, we need to use the result of this function
multiple times to apply the annotation and to determine to
send dark traffic. Given the function might be non-deterministic,
we break the API for ADTF to take a Boolean instead of a function.

JIRA Issues: CSL-11658

Differential Revision: https://phabricator.twitter.biz/D825317