finagle-core: add additional tracing counters
Problem:
We'd like to gather a better understanding of our
sampling rate and the discrepancies between the
rate we enforce and the actual number of traces we
see.

Solution:
To determine whether this is a problem on the
Finagle side or a problem due to Zipkin lossiness,
we'll add two additional counters along with the
`Tracing.sampled` metric: `Tracing.notSampled` and
`Tracing.active`.

Result:
These values should add up properly and should
give us insight into how well finagle is abiding
by the chosen sampling rate

JIRA Issues: CSL-11609

Differential Revision: https://phabricator.twitter.biz/D830444