finagle-core: Clarify the `.withMonitor(..)` docs
Problem

The doc string for `CommonParams.withMonitor(..)` are technically
incorrect in that the default monitor is actually the `NullMonitor`, not
the `DefaultMonitor`, but they're correct in spirit in that the
DefaultMonitor is composed with the configured monitor, and in the case
of the NullMonitor, that just leaves the DefaultMonitor. This can all be
very confusing.

Solution

Make the docs technically correct in that the NullMonitor is the default
Monitor configuration and also be more explicit that the configured
Monitor is composed with DefaultMonitor when added to the stack.

JIRA Issues: CSL-11213

Differential Revision: https://phabricator.twitter.biz/D831453