finagle-netty4-http: allow setting multiple values for a request cookie
Problem:
In a Response, `CookieMap` allows setting multiple cookie values for a given name
as long as they differ by their domain or path. However, in a Request, if the
client sent several values for a given cookie name, only the last one is
currently available. Request cookies do not have domain or path to distinguish
between multiple cookies with the same name.

Solution:
Fix a bug in `Netty4CookieCodec` when decoding Request cookies. Allow inserting
multiple values for a Request cookie for the same name in a CookieMap with
`add`,  `+=` and `addAll`.

Result:
On a Request, `CookieMap.add`, `CookieMap.+=` and `CookieMap.addAll` keep
existing cookies with the same name when a new cookie is added. On a Response,
the behavior is unchanged; multiple cookies with the same name, path, and domain
are deduplicated and only the last one is kept.

Closes twitter/finagle#911
Fixes twitter/finagle#909
Signed-off-by: Lily He <lhe@twitter.com>

JIRA Issues: CSL-11397

Differential Revision: https://phabricator.twitter.biz/D801466