finagle-stats-core: Fix broken MetricsHostStatsReceiverTest
Problem

We're querying for metrics using `Seq(name)` against a map keyed by a
simple `String`. This isn't caught by the compiler because it's a java
map and the `.get(..)` method only wants an `Object`. The tests happen
to work because, as it turns out, `null == null`.

Solution

Both fix the keys and make sure the keys we query for return non-null
values.

Differential Revision: https://phabricator.twitter.biz/D842563