finagle-core: Remove WhenNoNodesOpenParam from LoadBalancerFactory
Problem:
WhenNoNodesOpenParam is an unused stack param. No one configures it to a different
value than the default (PickOne).

Solution:
Remove `WhenNoNodesOpenParam`, `WhenNoNodesOpen`, `NoNodesOpenServiceFactory`, and
`NoNodesOpenException`. This removes the "fail fast" behavior where the service
factory throws an exception instead of trying to service the request.

Result:
When no nodes are Status.Open, the load balancer will pick one node to send the
request to. This is the current default.

JIRA Issues: CSL-11653

Differential Revision: https://phabricator.twitter.biz/D843176