finagle/finagle-netty4: Add ZStd streaming decompressor for use by clients
Problem/Solution:

Experiments last year showed possible performance gains for using ZStd compression rather than LZ4 (currently supported) for some of our customers. This change adds an implementation of ZStd compression (existing Netty implementation) and a new decompressor. If this proves fruitful, the intent is to upstream the decompressor. There is currently a non-functioning version in PR status with Netty and this could subsume it.

JIRA Issues: CSL-11598

Differential Revision: https://phabricator.twitter.biz/D840984