util-stats: Add ability to tell MetricBuilder that the dimensional name is not fully defined
Problem

We have a lot of metrics already defined in hierarchical form and we
need to add labels to them before they're meaningful in dimensional
form.

Solution

Add a new field to MetricBuilder that speaks to the validity of the
label state. This will also be helpful in the future for figuring out
what metrics are not fully labeled.

Differential Revision: https://phabricator.twitter.biz/D852745