finagle-stats-core: Refactor PrometheusExporter to allow easier configuration
Problem

We want to be flexible in how we write prometheus params but it turns
out there are a lot of different things that need to be threaded through
like verbosity, export metadata, should we export empty percentiles,
etc. This makes calling function signatures get messy.

Solution

Move all these functions into a class instead of an object and make
configuration arguments to the class, not functions. This makes them
available to all the (now) methods automatically.

We now take advantage of this through the exporter and thread
through both flags and query params.

Differential Revision: https://phabricator.twitter.biz/D866121