finagle-http: Add a retry for a racy connection pool
Problem

When testing MethodBuilder, we have a timeout, that cuts the connection.
However, sometimes the next request to that MethodBuilder runs into a
CloseChannelException.  Our suspicion is that the connection pool is racing
and a connection that will soon be closed is added to the connection pool.

Solution

We should deal with the race by just retrying.  If HTTP/1.1 was a more popular
protocol, it might be worthwhile to fix this, but since cutting the connection
on interrupt and using the default pool is an unusual combination in this day
and age, we decided to just adjust our tests.

JIRA Issues: CSL-11907

Differential Revision: https://phabricator.twitter.biz/D869592