finagle-mux: Make ClientServerTestDispatch less flaky
Problem:
The ClientServerTestDispatch "server nacks new requests after draining" test is occasionally flaky. This could be because MuxServerSession ensures server closure happens within a deadline of Time.now. If the deadline is reached, a ChannelClosedException is thrown for outstanding requests.

Solution:
Update the deadline to Time.Top to give the server unlimited time to drain. Also add an await to the final assertion to ensure the promise has time to be satisfied.

JIRA Issues: CSL-11889

Differential Revision: https://phabricator.twitter.biz/D869564