finagle-stats-core: Honor `UnlatchedCounters` in `CounterDeltas`
Problem

We want to be able to say that a metric is not latched even if the
global setting is set to allow latched metrics.

Solution

Inspect the type of the counter in `CounterDeltas` and if it's an
`UnlatchedCounter` type just emit the snapshot value as is.

JIRA Issues: CSL-11887

Differential Revision: https://phabricator.twitter.biz/D869424