finagle-thrift: Make EndToEndTest less flaky
Problem:

The finagle-thrift EndToEndTest test is occasionally flaky. It fails on the assertion that the total request count is 5, when sometimes it is 6. This could be happening because of retries since the stats filter is below the retry filter in the client stack.

Solution:

Move the stats filter to the top of the stack for these tests to ensure it only counts logical requests, and doesn't count retries.

JIRA Issues: CSL-11910

Differential Revision: https://phabricator.twitter.biz/D870418