finagle/finagle-stats-core: Change filter file processing to accept non-existent file paths and instead treat them as empty
Problem/Solution

Passing an incorrect value for a metrics filter file caused no metrics to be emitted at all.
The desired behavior is to absorb this exception and treat it as an empty filter.

The code already logs the missing file but now instead of rethrowing, it treats it the same as an empty file.
The logging level has been reduced to WARN and the message makes clear that the process is ignoring it and continuing.

JIRA Issues: CSL-11311

Differential Revision: https://phabricator.twitter.biz/D879237