finagle/finagle-core: Add flag to cull unreachable branches in a NameTree.Union
Problem/Solution:

When a user constructs a NameTree.Union of weighted subsets, the default behavior in
NameTreeFactory is to retain any subset that has a weight > 0, even if that subset is unreachable.
For example, in a Union of (X% NameTree.Empty, Y% NameTree.Leaf()), X% of calls would fail
as the load balancer would select the unreachable subset X% of the time.

This solution adds a global flag which can change this behavior to cull those subsets while
maintaining the default of today's logic so that anybody who either does this on purpose or
uses the behavior for something else will be unaffected.

JIRA Issues: CSL-11952

Differential Revision: https://phabricator.twitter.biz/D879306