finagle/finagle-benchmark: finagle microbenchmark StringServer & StringClient
Problem/solution

We want to see how Finagle performs when we remove all of the
extra features from the server-side and client-side. For the Server side,
It's inserted after ServerTracingFilter, since it's at the very bottom of the stack.
Meaning echoStringModule will be hit after all other Finagle modules/filters in
the stack (instead of netty). For the Client side,  we have a full client until the
load balancer layer of the stack.

JIRA Issues: CSL-11870

Differential Revision: https://phabricator.twitter.biz/D853163