finagle-core: Fix a race condition where a coord update can get dropped in Balancer
Problem

We have a race condition that affects the deterministic aperture
balancers. This race can happen when a peerset update races with other
rebuild events and a peerset update could be missed for a cycle.

Solution

Make the `def rebuild()` method actually do a rebuild unconditionally.
In the process we switch to a `ReentrantLock` instead of synchronization
so that we can avoid blocking for non-mandatory rebuilds caused by
load balancer panics. This means that only updates and calls to `rebuild()`
may block the thread.

Differential Revision: https://phabricator.twitter.biz/D829153