util-stats: More reshaping of dimensional metrics support
Problem

Recent plans have made scopes more valuable for us and less things will
likely be changed from scopes to labels. This is great news as that
means most scopes are still valid. However, we need to account for that
in our model which doesn't allow scopes for dimensional metrics.

Solution

Simply allow it. We change the identity to be the union of the
hierarchical and dimensional names plus the labels and whether the metric
is valid dimensionally.

JIRA Issues: CSL-11965

Differential Revision: https://phabricator.twitter.biz/D891342