finagle-core: Make the c.t.f.l.aperture.minDeterministicAperture flag less temporary
Problem

The minDeterministicAperture flag was intended to be temporary while we
figured out what the real value should be. However, we've found out that
there is not perfect value for everyone so while 12 is a good value for
most some people do want to tune it.

Solution

Give it a doc string instead of a comment and add an entry to the docs
for it.

Differential Revision: https://phabricator.twitter.biz/D905924