finagle-core: Remove the "eager_connections" metric
Problem

We are emitting an "eager_connections" metric to make it easier for
service owners to know if their load balancers are set to use eager
connections. This was helpful during the rollout as it was based on a
toggle, at the time. However, now eager connections is fully rolled out
so the whole cluster will have the same behavior and the metric is less
valuable.

Solution

Remove the metric and instead store the info in the balancer metadata so
it can be accessed from the admin page, if necessary.

Differential Revision: https://phabricator.twitter.biz/D905836