finagle-core: Cleanup some code in StatsFilter
Problem

We have two different types in the StatsFilter.scala file: both the
actual StatsFilter and the StatsServiceFactory which are kind of along
similar lines but not actually directly related.

Solution

Split the StatsServiceFactory out into it's own file. While were there
we can also eagerly remove the gauge on the close of the ServiceFactory.

Differential Revision: https://phabricator.twitter.biz/D907483