finagle-stats: Treat negative values as 0s in histograms
Problem

There is inconsistency in how we treat negative values in our histograms. For the most part,
they are treated as 0s with a small exception: in `average` and `sum` summary they are treated
as negative values. This leaves our histograms in an inconsistent state.

Solution

Truly treat negative values as 0s in our histograms.

JIRA Issues: CSL-12050

Differential Revision: https://phabricator.twitter.biz/D907434