finagle-core: Add source info in Exceptions that can be propagated
Problem:

`ClientDiscardedRequestException`s sometimes can wrap exceptions that are
propagated multiple levels.

There are mainly three kinds of exceptions that will raise CDREs:
1. c.t.finagle.TimeoutException and its successors
2. Failures that are thrown by BackupRequestFilter
3. c.t.util.TimtoutException

Finagle TimeoutExceptions only expose the client label, and it's not clear on
which server/job that client exists.
BackupRequest Failure is initiated statically with noSource.

Solution:

Record server identifier in ExceptionSourcedFilter and expose it in TimeoutException.
Add the Failure.Source.AppId and fill it in for the first two exceptions

Result:

BackupRequestFilter Failures and Finagle TimeoutExceptions carry server info when
it's available.

Not sure anything we can do for util TimeoutExceptions

JIRA Issues: CSL-11913

Differential Revision: https://phabricator.twitter.biz/D886963