finagle-core: Cleanup some tests for StandardStatsReceiver
Problem

We're going to change the StandardStatsReceiver in the near future to
support dimensional metrics but it's a bit tough right now because it
depends on global state.

Solution

Make the counter plugable but default to the global state so it's easy
to inject new state for testing. Also add a test that the request
latency metric is there.

Differential Revision: https://phabricator.twitter.biz/D914795