finagle-memcached: Add option to create a twemcache client without partitioning service inserted
Problem

The Memcached client only allows users to create a client that uses the partitioning service. We are going to introduce a stateless service in front of cache clusters so clients will we able to use a cache cluster like any other stateless service.

Solution

Add a new function that creates a Memcached client that doesn't insert the partitioning service

Differential Revision: https://phabricator.twitter.biz/D911789