finagle-core: Extract standard metrics out of StatsFilter
Problem

We have a bunch of metrics that we create in StatsFilter, and they're basically
duplicates with different names and classification rules. We're already partially
extracting them out into the RPCMetrics case class, but it's still in StatsFilter
and a bit cluttered.

Solution

Go all the way and extract the metrics into their own class, BasicServiceMetrics.
This gives us a central place to think about these metrics and the StatsFilter is now
just in charge of letting BasicServiceMetrics know that a request happened.

Differential Revision: https://phabricator.twitter.biz/D917812