finagle-core: Only sample the timestamp when needed
Problem

Zipkin Thrift only expects timestamps in standard annotations, but we record
precise timestamps for all annotations, both standard and binary. This is
in large part an implementation detail--in finagle-core, binary annotations
are a kind of annotation, and they're all recorded with a timestamp.

Solution

Have a separate method for handling binary annotations that don't need a
timestamp so that we avoid the overhead of doing a system call to grab a
timestamp.

JIRA Issues: CSL-11964

Differential Revision: https://phabricator.twitter.biz/D883285