finagle-core: add tracing to DeadlineFilter
Problem:
Customers want to be able to tell if the deadline
has been modified as it passes to downstreams, as well
as where it originates.

Solution:
Add four new trace annotations to DeadlineFilter:
`<clnt/srv>/request_deadline` shows the time at which the
deadline will be reached
`<clnt/srv>/request_deadline_remaining_ms` shows the time
remaining before the deadline is reached. This appears
only when the deadline has not yet been reached.
`<clnt/srv>/request_deadline_exceeded_ms` shows in milliseconds
how much we've exceeded the deadline. This appears only
when the deadline has been exceeded.
`<clnt/srv>/request_deadline_rejected` is true when a request has
been rejected because there are enough tokens to reject it.
It appears only when a request has been rejected.

Note that to get the appropriate scoping, we have created a
clientModule and serverModule

JIRA Issues: CSL-12203

Differential Revision: https://phabricator.twitter.biz/D943426