finagle-mux/finagle-netty: Add ZStd compression support for mux
Problem:

The Netty provided implementation of the ZStd compressor sends each message as a
frame. This causes a lot of churn on the decompression side as the setup and
teardown of the decompression pipeline is expensive. We have empirical evidence
for this through ads-ranking/prediction testing using shadow traffic. The effect
is massive losses to GC pauses as the decompressor builds up and disposes native
handles.

Solution:

We want our own implementation that takes advantage of our long lived
connections that keeps a frame open, sending messages as blocks instead of
frames.

JIRA Issues: CSL-11598

Differential Revision: https://phabricator.twitter.biz/D868248