finagle-core: add sample rate flag
Problem:
Currently, the sample rate is only written
to the root trace. This makes analysis and
debugging difficult, as it requires manual
backtracing to determine the sampling rate.

Solution:
As part one to solving this problem, we
introduce the `setSampleRate` and `getSampleRate`
APIs to our tracing flags. These API's use the
32 bits succeeding the `Debug`, `SamplingKnown`, and
`Sampled` binary flags to encode the float value of
the sample rate.

JIRA Issues: CSL-12204

Differential Revision: https://phabricator.twitter.biz/D952121