finagle-http: Allow for multipart FileElement without Content-Transfer-Encoding: binary
Problem

We hard-code all FileElements as binary so an appropriate `Content-Transfer-Encoding` head is
added for them. In some cases, we don't want to set any `Content-Transfer-Encoding` (i.e.,
when FileElement is text/plain).

Solution

Expose `isText` attribute for FileElement (default is `false`, which doesn't change today's behavior).
When set to `true`, no `Content-Transfer-Encoding` header is added to the request.

Differential Revision: https://phabricator.twitter.biz/D969859