finagle-core: Enable post-partitioning backup requests
Problem

When backup requests are enabled on a partitioned client, the backups are sent
for each logical request, resulting in a lot of additional physical requests if
the client is fanning out the request to all of its partitions. For example, if
a client is fanning out one request to 10 partitions and one of those partitions
is slow, the backup request will be initiated above the partitioning client and
each partition will receive a backup request.

Solution/Result

Change the position of the BackupRequestFilter in the client stack so that it
comes after the partitioning client so that when we are sending backups, the
backups are calculated on the physical requests sent to the partitions not the
logical requests. This should improve latencies when waiting for a response from
a partitioned request where some of the partitions are slow.

JIRA Issues: CSL-12035

Differential Revision: https://phabricator.twitter.biz/D930610