finagle: Mask interrupts on connection's rollback (and/or poison)
Problem / Solution

When any MySql connection is closed (put back in the pool), a Finagle client will
try to "rollback" (to deal with MySql's commands that cause implicit commit) it to
the state where can be safely reused later. Sometimes an interrupt may come in at
the same time the client is running a rollback, which leads to a connection being
invalidated and thrown away.

It's probably a wasteful thing to do given that caller has already signaled that it's
no longer interested in any data coming from this session. That is, if we're able
to successfully roll it back, we might be able to reuse it.

Solution

Just like we do in transactions, "mask" interrupts on the rollback command that's
run at the end of each session.

Differential Revision: https://phabricator.twitter.biz/D968787