finagle-core: Undo DynamicBRF for HTTP clients
Problem / Solution

There is no need for `DynamicBRF` to be installed in HTTP clients as they don't support partitioning.
Let's remove it.

We also don't need to install `DynamicBRF` eagerly in ThrftMux clients unless they are used with
MethodBui Let's install the placeholder (no-op) instead and replace it in MethodBuilder.

Differential Revision: https://phabricator.twitter.biz/D979376