finagle-http2: fix tests ignored after mockito upgrade
Problem:
The recent upgrade from mockito-all to mockito-core exposed a problem
in our testing logic for `PriorKnowledgeHandlerTest`. When getting the next
handler in a pipeline, netty [[ https://github.com/netty/netty/blob/4.1/transport/src/main/java/io/netty/channel/AbstractChannelHandlerContext.java#L1052-L1058 | skips over ]] handlers that are unnecessary. The
`ChannelInboundHandlerAdapter` we attempt to mock and track ends up getting
skipped because all of its methods are marked @ Skip, which results in no
`channelRead` call. In mockito-all, the mock somehow changed the object enough to
overcome this, but that has since been fixed in mockito-core

Solution:
Create a `TestHandler` that extends the `ChannelInboundHandlerAdapter` and
overrides `channelRead` with the same behavior as its parent. This ensures that at
least one method, the one we're hoping to call, is not marked skippable and
therefore the handler is not skipped.

JIRA Issues: CSL-12361

Differential Revision: https://phabricator.twitter.biz/D997874