finagle-core: introduce FanoutTracer
Problem:
When a single Memcached
client request turns into multiple requests to
individual cache shards via partitioning, we
still only get a single span in the zipkin traces

Solution:
Along with moving the `MemcachedTracingFilter`
below the `MemcachedPartitioningService` in the
stack, this solution also introduces
two new features to finagle's tracing
infrastructure. First, the `FanoutTracer`, which
wraps a tracer and buffers `record(r)` calls until
the traces are flushed via `flush()`. Second,
introduce a `Trace.peerId` API, which behaves
similarly to `Trace.nextId`, but makes the `parentId`
for the new trace the same as that of the current
trace (as opposed to making the `parentId` of the new
trace the `spanId` of the current). Finally, the `let[Fanout]Tracer`
APIs, which ensure that all Tracers are wrapped in
FanoutTracers if fanout is enabled. This ensures that
no individual Tracer flushes pre-partitioning records to
the pre-partitioning TID.

In tandem, these three features accomplish two goals:
get us unique peer spans for fanout request traces, and
ensure that each those spans are populated with the
annotations recorded before partitioning in the stack.

JIRA Issues: CSL-10079, CSL-12097

Differential Revision: https://phabricator.twitter.biz/D974564