finagle-serverset2: remove Endpoints with a weight of -1.0 from serverset for testing
Problem:
We would like to use a negative weight of -1.0 to signal the removal of an Endpoint from a resolved serverset that is returned to a client.
When the weight is >= 0 (or the Vector removed), it will be included in the serverset again.
This can be a useful tool for certain kinds of load testing, specifically Squeeze Testing where we wish to remove serverset members at a measured rate to induce increased load on the remaining members.

Solution:
In Zk2Resolver, filter out Endpoints with a weight of -1.0 when we validate them while resolving their raw address.

Note:
When using deterministic Aperture is important to remove nodes in descending order or the load balancer will "save" a place for the missing node and not rebalance the ring which can lead to non-optimal traffic patterns.

JIRA Issues: CCS-57

Differential Revision: https://phabricator.twitter.biz/D1001342