finagle-netty4: Cleanup SharedChannelStats
Problem

There are two main problems:
1. We're exposing the LongAdder in SharedChannelStats as a protected field
and also with methods that increment/decrement it. We really just need
the latter and exposing the LongAdder is exposing unnecessary
implementation details.
2. The eventloop pending event counting is broken: we add and remove loops
to our set on channel connect and disconnect events, but the loops themselves
are shared among many channels. This means that our connects are fine because
we de-dupe due to the set, but on disconnect we may remove an eventloop that
is used by more than one channel which is likely to be the common occurrence.

Solution

Just make the LongAdder private and include connects increments in the method.
We also simply use the EventLoopGroup specified in the params, which is the
same one that the client/server will use, and iterate over it's members to
get the pending io events.

Differential Revision: https://phabricator.twitter.biz/D986493