finagle-serversets: remove weight gauge
Problem:
The `weightRemovals` gauge currently overcounts the number of entries with a
weight of -1.

Solution:
Simply remove the gauge.

Reason:
Setting the weight of an instance to -1 is a relatively infrequent operation.
Therefore, not every resolution needs a gauge to track the number of filtered
entries. For the subset of services which utilize negative weights, the number
of filtered entries can be deduced by comparing the number of member nodes in ZK
to the `size` gauge.

Result:
One less gauge per resolution.

Differential Revision: https://phabricator.twitter.biz/D1004280