util-stats: Update HistogramComponent.DefaultPercentiles to use correct values & respect metrics implementation
### Problem
The motivation for the phab is to fix the bug where the `HistogramComponent.DefaultPercentiles` is set to incorrect values (i.e. 99 as opposed to 0.99), resulting in additional trailing zeroes in the metric name. In the initial review, there was a concern around how different metrics implementation (i.e. finagle vs ostrich) may need to handle the input differently.

### Solution

1. Fix the input passed to HistogramComponent.DefaultPercentiles
1. Update `ExpressionSchema` to respect different types of metrics implementation. `StatsFormatter` already handles the differences in metrics implementation, so this change includes a refactor so that `StatsFormatter` and `HistogramFormatter` share the implementation for generating the correct labels.

JIRA Issues: CSL-12353, SSTR-7425

Differential Revision: https://phabricator.twitter.biz/D999055