finagle-core: Support service-loading the offload filters thread pool
Problem

It's currently difficult to change the default thread pool used by
offload filter because of how it's constructed. This makes it difficult
to add custom behavior and telemetry to the thread pool.

Solution

Allow a service loaded pathway so that users can add custom thread
pools.

JIRA Issues: CSL-12330

Differential Revision: https://phabricator.twitter.biz/D1011816