[finagle] do not fail full toggle map parsing when duplicate toggles exist in a file
Problem:

Currently the finagle toggle map parser fails parsing of all elements in a
toggle file when a single duplicate entry is added. While this fail fast
behavior is good in some cases (especially when toggles are loaded locally),
it's not a great behavior in production systems where toggles are committed and
loaded separately. In such scenarios, if the file is already committed and old
version is unavailable to the system it is preferable to have the toggle map
load as much as it can to prevent breaking other toggles in the file.

Solution:

Add an opt in setting to keep the first entry when a duplicate is encountered
instead of failing to parse the whole thing

Differential Revision: https://phabricator.twitter.biz/D1117265