finagle-memcached: Implement compressing cache using lz4 compression
**Problem** Make lz4 compression easily available via supported memcache
clients. The desired effect is to drastically reduce the size of blobs stored in
memcache, with minimal CPU cost to the calling services.

**Solution** The hypothesis is that we can drastically reduce the size of our
typical memcache payloads, at very low CPU cost, by using a throughput-oriented
compressor like lz4. This will drastically reduce the network utilization of
caches, improving performance in high-load situations like site failovers. This
will also improve the storage utilization of cache, which will reduce eviction
(and perhaps increase cache hit rates of calling services). For larger blobs
that span >1 packet, this might also reduce cache tail latencies, by reducing
the number of packets needed to transmit an object and thereby reducing the
chances of any one of them being delayed or needing a retransmit.

By using a few bits in the memcache protocol's flags field, this can be
implemented in a way that allows for a transparent upgrade. New blobs written to
the store will be compressed - existing blobs will be unmodified. The bits in
the flags field will be used to signal whether the blob is compressed.
Transparent upgrades will make adoption of this significantly easier, as the
existing cached data in the cluster can be used while the transition occurs.

**Result** ~30% compression for the higher sized items and 15-20% from p40
onwards

Differential Revision: https://phabricator.twitter.biz/D1130236